Move plan util from common into lib utils
