+++ enterprise/frontend/src/metabase-enterprise/embedding/selectors.ts
@@ -1,6 +1,6 @@
-import { getPlan } from "metabase/common/utils/plan";
 import type { State } from "metabase/redux/store";
 import { getSetting } from "metabase/settings";
+import { getPlan } from "metabase/utils/plan";
 
 export const isInteractiveEmbeddingEnabled = (state: State) => {
   const plan = getPlan(getSetting(state, "token-features"));
