+++ enterprise/frontend/src/metabase-enterprise/security_center/components/SecurityCenterBanner/SecurityCenterBanner.tsx
@@ -6,9 +6,9 @@ import {
 } from "metabase/api";
 import { Banner } from "metabase/common/components/Banner";
 import { Link } from "metabase/common/components/Link";
-import { getPlan } from "metabase/common/utils/plan";
 import { useSetting } from "metabase/settings";
 import { Anchor, Text } from "metabase/ui";
+import { getPlan } from "metabase/utils/plan";
 
 import { isAffected } from "../../utils";
