+++ enterprise/frontend/src/metabase-enterprise/security_center/components/SecurityCenterPromoCard/SecurityCenterPromoCard.tsx
@@ -6,12 +6,13 @@ import {
   useGetChannelInfoQuery,
   useListSecurityAdvisoriesQuery,
 } from "metabase/api";
-import { getPlan } from "metabase/common/utils/plan";
 import { getUserIsAdmin } from "metabase/current-user";
+import { getPlan } from "metabase/utils/plan";
 import { NavbarPromoCard } from "metabase/nav/components/NavbarPromoCard";
 import { useSelector } from "metabase/redux";
 import { useSetting } from "metabase/settings";
 import { Icon } from "metabase/ui";
+import { getPlan } from "metabase/utils/plan";
 
 import { isAffected } from "../../utils";
