+++ enterprise/frontend/src/metabase-enterprise/semantic_search/SearchSettingsWidget.tsx
@@ -6,11 +6,11 @@ import { SettingHeader } from "metabase/admin/settings/components/SettingHeader"
 import { BasicAdminSettingInput } from "metabase/admin/settings/components/widgets/AdminSettingInput";
 import { UpsellSemanticSearchPill } from "metabase/admin/upsells/UpsellSemanticSearch";
 import { getErrorMessage } from "metabase/api/utils";
-import { getPlan, isProPlan } from "metabase/common/utils/plan";
 import type { SearchSettingsWidgetProps } from "metabase/plugins";
 import { useSelector } from "metabase/redux";
 import { getSetting, useAdminSetting } from "metabase/settings";
 import { Box, Progress, Stack, Text, Tooltip } from "metabase/ui";
+import { getPlan, isProPlan } from "metabase/utils/plan";
 import { useGetSemanticSearchStatusQuery } from "metabase-enterprise/api/search";
 
 function useLatch(bool: boolean) {
