+++ frontend/src/metabase/admin/people/selectors.ts
@@ -1,10 +1,11 @@
 import { createSelector } from "@reduxjs/toolkit";
 
 import { ACTIVE_USERS_NUDGE_THRESHOLD } from "metabase/admin/people/constants";
-import { hasAnySsoFeature } from "metabase/common/utils/plan";
 import { getUserIsAdmin } from "metabase/current-user";
+import { hasAnySsoFeature } from "metabase/utils/plan";
 import type { State } from "metabase/redux/store";
 import { getSetting } from "metabase/settings";
+import { hasAnySsoFeature } from "metabase/utils/plan";
 
 export const getUserTemporaryPassword = (
   state: State,
