+++ frontend/src/metabase/admin/permissions/selectors/data-permissions/permission-editor.tsx
@@ -8,7 +8,6 @@ import {
   getSpecialGroupType,
   isDefaultGroup,
 } from "metabase/common/utils/groups";
-import { getPlan } from "metabase/common/utils/plan";
 import {
   PLUGIN_AUDIT,
   PLUGIN_FEATURE_LEVEL_PERMISSIONS,
@@ -18,6 +17,7 @@ import type { State } from "metabase/redux/store";
 import { getMetadataWithHiddenTables } from "metabase/selectors/metadata";
 import { getSetting, getTokenFeature } from "metabase/settings";
 import { getResponseErrorMessage } from "metabase/utils/errors";
+import { getPlan } from "metabase/utils/plan";
 import type Schema from "metabase-lib/v1/metadata/Schema";
 import type {
   Database,
