+++ frontend/src/metabase/admin/settings/components/CloudPanel/MigrationSuccess.tsx
@@ -1,9 +1,9 @@
 import { t } from "ttag";
 
 import { ExternalLink } from "metabase/common/components/ExternalLink";
-import type { Plan } from "metabase/common/utils/plan";
 import { Box, Button, Flex, Icon, Text } from "metabase/ui";
 import { color } from "metabase/ui/colors";
+import type { Plan } from "metabase/utils/plan";
 import type { CloudMigration } from "metabase-types/api/cloud-migration";
 
 import { LargeIconContainer, MigrationCard } from "./CloudPanel.styled";
