+++ frontend/src/metabase/admin/settings/components/CloudPanel/utils.ts
@@ -1,5 +1,5 @@
-import type { Plan } from "metabase/common/utils/plan";
 import { dayjs } from "metabase/dayjs";
+import type { Plan } from "metabase/utils/plan";
 import type {
   CloudMigration,
   CloudMigrationState,
