+++ frontend/src/metabase/admin/upsells/UpsellRemoteSync.tsx
@@ -5,11 +5,11 @@ import { UpsellBigCard } from "metabase/common/components/upsells/components";
 import S from "metabase/common/components/upsells/components/Upsells.module.css";
 import { UPGRADE_URL } from "metabase/common/components/upsells/constants";
 import { useHasTokenFeature } from "metabase/common/hooks";
-import { getPlan, isProPlan } from "metabase/common/utils/plan";
 import { PLUGIN_ADMIN_SETTINGS } from "metabase/plugins";
 import { useSelector } from "metabase/redux";
 import { getApplicationName } from "metabase/selectors/whitelabel";
 import { useSetting } from "metabase/settings";
+import { getPlan, isProPlan } from "metabase/utils/plan";
 
 export const UpsellRemoteSync = ({ source }: { source: string }) => {
   const hasRemoteSync = useHasTokenFeature("remote_sync");
