+++ frontend/src/metabase/home/components/EmbedHomepage/EmbedHomepage.tsx
@@ -2,12 +2,12 @@ import { useState } from "react";
 import { t } from "ttag";
 
 import { useHasTokenFeature } from "metabase/common/hooks";
-import { getPlan } from "metabase/common/utils/plan";
 import { PLUGIN_IS_EE_BUILD } from "metabase/plugins";
 import { useDispatch, useSelector } from "metabase/redux";
 import { addUndo } from "metabase/redux/undo";
 import { getDocsUrl } from "metabase/selectors/settings";
 import { getSetting, useSetting } from "metabase/settings";
+import { getPlan } from "metabase/utils/plan";
 import type { EmbeddingHomepageDismissReason } from "metabase-types/api";
 
 import { useSendProductFeedbackMutation } from "../../api/product-feedback";
