+++ frontend/src/metabase/selectors/settings.ts
@@ -1,8 +1,8 @@
 import { createSelector } from "@reduxjs/toolkit";
 
-import { getPlan } from "metabase/common/utils/plan";
 import type { State } from "metabase/redux/store";
 import { getSetting } from "metabase/settings";
+import { getPlan } from "metabase/utils/plan";
 import type { TokenStatus, Version } from "metabase-types/api";
 
 export const isSsoEnabled = (state: State) =>
