+++ frontend/src/metabase/transforms/pages/EnableTransformsPage/EnableTransformsCard.tsx
@@ -2,7 +2,6 @@ import { jt, t } from "ttag";
 
 import { useListDatabasesQuery } from "metabase/api/database";
 import { Link } from "metabase/common/components/Link";
-import { getPlan, isProPlan } from "metabase/common/utils/plan";
 import { useSelector } from "metabase/redux";
 import { getSetting } from "metabase/settings";
 import { doesDatabaseSupportTransforms } from "metabase/transforms/utils";
@@ -17,6 +16,7 @@ import {
   Text,
   Title,
 } from "metabase/ui";
+import { getPlan, isProPlan } from "metabase/utils/plan";
 import type { IconName } from "metabase-types/api";
 export function EnableTransformsCard({
   onEnableClick,
